Odradjena autorizacija na par kontrolera i dodata provera vec postojeceg termina

+++ SpringWebApp/src/main/java/isamrs/controller/LekController.java
@@ -3,6 +3,8 @@ package isamrs.controller;
 import java.util.Collection;
 import java.util.stream.Collectors;
 
+import isamrs.domain.AdministratorKlinickogCentra;
+import isamrs.domain.Lekar;
 import isamrs.domain.Recepti;
 import isamrs.dto.LekDTO;
 import isamrs.exceptions.NotFoundException;
@@ -24,6 +26,8 @@ import org.springframework.web.server.ResponseStatusException;
 import isamrs.domain.Lek;
 import isamrs.service.LekService;
 
+import javax.servlet.http.HttpServletRequest;
+
 @RestController
 @RequestMapping("/lek")
 public class LekController {
@@ -35,14 +39,20 @@ public class LekController {
 	private ReceptService recService;
 	
 	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Collection<LekDTO>> getLekovi(){
+	public ResponseEntity<Collection<LekDTO>> getLekovi(HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)&&!(request.getSession().getAttribute("user") instanceof Lekar)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		Collection<Lek> lekovi = service.findAll();
 		Collection<LekDTO> lekDTOS = lekovi.stream().map(this::LektoDTO).collect(Collectors.toList());
 		return new ResponseEntity<Collection<LekDTO>>(lekDTOS, HttpStatus.OK);
 	}
 	
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<LekDTO> getLek(@PathVariable("id") Long id){
+	public ResponseEntity<LekDTO> getLek(@PathVariable("id") Long id, HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra )){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		try {			
 			Lek l = service.findOne(id);
 			return new ResponseEntity<LekDTO>(LektoDTO(l), HttpStatus.OK);
@@ -53,19 +63,28 @@ public class LekController {
 	}
 	
 	@PostMapping(produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<LekDTO> createLek(@RequestBody Lek l){
+	public ResponseEntity<LekDTO> createLek(@RequestBody Lek l, HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		Lek l2 = service.create(l);
 		return new ResponseEntity<LekDTO>(LektoDTO(l2), HttpStatus.CREATED);
 	}
 
 	@PutMapping(value = "/{id}",produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<LekDTO> updateLek(@RequestBody Lek l, @PathVariable("id") Long id) throws NotFoundException {
+	public ResponseEntity<LekDTO> updateLek(@RequestBody Lek l, @PathVariable("id") Long id, HttpServletRequest request) throws NotFoundException {
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		Lek l2 = service.update(id, l);
 		return new ResponseEntity<LekDTO>(LektoDTO(l2), HttpStatus.OK);
 	}
 	
 	@DeleteMapping(value = "/{id}")
-	public ResponseEntity<LekDTO> deleteLek(@PathVariable("id") Long id){
+	public ResponseEntity<LekDTO> deleteLek(@PathVariable("id") Long id, HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		service.delete(id);
 		return new ResponseEntity<LekDTO>(HttpStatus.NO_CONTENT);
 	}
